migrate(execution-policy): extract the policy decision from agent_policy.c (Class A #4)
Fourth Class A migration. The tool-action policy decision — policy_load (reads the operator
.aimee-policy.json) and policy_check_tool (the fail-closed allow/deny decision), plus the g_policy_json
static, the contiguous "Policy checking" section of src/server/agent_policy.c — is extracted into
src/modules/execution-policy/execution_policy.c and the descriptor latches. 22/26 latched; four Class A
modules remain.

Boundary (self-reviewed against docs/modules/execution-policy.md): the module owns the decision, not
validation. A first broad extraction that pulled tool_validate and the schema-validation helpers into
execution-policy was rejected as doc-violating ("schema and argument validation such as tool_validate
stays with tools") and redone decision-only:
- moved: policy_load, policy_check_tool, g_policy_json;
- stayed in agent_policy.c: tool_validate/normalize_args/validate_against_schema/tool_suggest and the
  tool-prompt helpers, tool_side_effect, and the trace/metrics/manifest half.

Link-clean by construction: g_policy_json is used only in the moved section, the moved functions call
no staying function, and nothing staying calls policy_load/policy_check_tool (external callers use the
shared headers/agent_exec.h). The decision contract stays in agent_exec.h, which the module implements
while the server implements the rest (the memory/DB2 arrangement), so execution-policy has no
module-private header. Guardrails pre_tool_check and gateway policing stay as decision consumers.

execution_policy.c is added to every source/object/test-link list carrying agent_policy.c/.o;
agent_policy.o keeps its tool_prompts_data.h rule, execution_policy.o uses the generic rule. tests is
empty (no dedicated policy-decision test; benchmarks precedent). Docs/register/ledger/test-suite updated.

Verified: all module/descriptor checks + the descriptor suite pass; nm confirms execution_policy.o
defines policy_check_tool and not tool_validate/trace; aimee and aimee-server build and link. Full
Make + CMake .253 validation to follow. The linker enforces the partition.
