+++ src/modules/execution-policy/module.yaml
@@ -8,5 +8,14 @@
   ],
   "runtime_toggle": {
     "supported": false
-  }
+  },
+  "ownership_complete": true,
+  "sources": [
+    "src/modules/execution-policy/execution_policy.c"
+  ],
+  "private_headers": [],
+  "tests": [],
+  "docs": [
+    "docs/modules/execution-policy.md"
+  ]
 }
